requirements.txt
asgiref==3.2.10
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
importlib-metadata==2.0.0
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
zipp==3.2.0
django-cors-headers==3.4.0
mysqlclient==1.4.6
djangorestframework-simplejwt==4.4.0
## The following requirements were added by pip freeze:
asn1crypto==1.4.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==2.9.2
docutils==0.15.2
idna==2.9
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.19
oauthlib==3.1.0
oscrypto==1.2.1
pycparser==2.20
pycryptodomex==3.9.8
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.3.2
urllib3==1.25.10

celery==5.0.0
redis==3.5.3
# celery[redis]==5.0.0
django-dotenv==1.4.2